requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==3.2
django-debug-toolbar==3.4.0
exceptiongroup==1.0.4
flake8==4.0.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pkg_resources==0.0.0
pluggy==1.0.0
psycopg2-binary==2.9
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
